--- src/components/login/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginButton } from './LoginButton';
